move ProximityPreconditions to its package
